refactor jsapi - configure
