+++ src/jsapi/configure.cc
@@ -1,4 +1,4 @@
-#include "configure.h"
+#include "include/configure.h"
 
 #include "process_data.h"
 #include "util-inl.h"
