+++ src/jsapi/http.cc
@@ -1,4 +1,4 @@
-#include "http.h"
+#include "include/http.h"
 
 #include "environment_data.h"
 #include "logger.h"
